fixed hide and show functionality of eoc collection

+++ chrome/content/zotero/xpcom/expressionsOfConcern.js
@@ -65,7 +65,7 @@ Zotero.ExpressionsOfConcern = {
 		 *        - look into the dom
 		 *        - extract expressions of concerns
 		 */
-		const items = await this.lookupExpressionsOfConcernItems();
+		const items = await this.lookupUrlsForEocs();
 		if (items) {
 			this.scrapeExpressionsOfConcern(items);
 		}
@@ -136,16 +136,33 @@ Zotero.ExpressionsOfConcern = {
 
 	/**
 	 *
-	 * @param itemID {string | number} primary key of the item which has an expression of concern
+	 * @param itemID {Zotero.Item} primary key of the item which has an expression of concern
 	 * @param data {{shortTexts: *, links: *}} data which gets stored into the expression of concern
 	 * @returns {Promise<void>}
 	 * @private
 	 */
-	_addEntry: async function (itemID, data) {
+	_addEntry: async function (item, data) {
 		const queryString = "INSERT OR IGNORE INTO expressionsOfConcern (itemID, data) VALUES (?, ?)";
-		this._expressionsOfConcern.set(itemID, this.FLAG.NORMAL);
-		this._queuedItemIDs.add(itemID);
-		await Zotero.DB.queryAsync(queryString, [itemID, JSON.stringify(data)]);
+		this._expressionsOfConcern.set(item.id, this.FLAG.NORMAL);
+		this._queuedItemIDs.add(item.id);
+		await Zotero.DB.queryAsync(queryString, [item.id, JSON.stringify(data)]);
+
+		let flag = this._expressionsOfConcern.get(item.id);
+		let libraryID = item.libraryID;
+
+		if (flag === undefined) {
+			this._expressionsOfConcern.set(item.id, this.FLAG.NORMAL);
+		}
+
+		if (!item.deleted && flag !== this.FLAG.HIDDEN) {
+			if (!this._expressionsOfConcernByLibrary[libraryID]) {
+				this._expressionsOfConcernByLibrary[libraryID] = new Set();
+			}
+			this._expressionsOfConcernByLibrary[libraryID].add(item.id);
+			await this._updateLibraryExpressionsOfConcern(libraryID);
+		}
+
+		await Zotero.Notifier.trigger('refresh', 'item', [item.id]);
 	},
 
 
@@ -200,14 +217,19 @@ Zotero.ExpressionsOfConcern = {
 	 * @private
 	 */
 	_removeEntry: async function (itemID, libraryID) {
-		this._expressionsOfConcern.delete(itemID);
-		this._expressionsOfConcernByLibrary[libraryID].delete(itemID);
-		this._updateLibraryExpressionsOfConcern(libraryID);
+		if (!this._expressionsOfConcern.has(itemID)) {
+			return;
+		}
 
 		const queryString = "DELETE FROM expressionsOfConcern WHERE itemID=?";
 		await Zotero.DB.queryAsync(queryString, itemID);
 
-		await Zotero.Notifier.trigger("trash", "expressionOfConcern", [itemID]);
+		this._expressionsOfConcern.delete(itemID);
+		this._expressionsOfConcernByLibrary[libraryID].delete(itemID);
+		this._updateLibraryExpressionsOfConcern(libraryID);
+
+
+		await Zotero.Notifier.trigger("refresh", "item", [itemID]);
 	},
 
 	/**
@@ -236,7 +258,7 @@ Zotero.ExpressionsOfConcern = {
 		await Zotero.Notifier.trigger('modify', 'item', [itemID]);
 	},
 
-	checkQueuedItemsInternal: async function() {
+	checkQueuedItemsInternal: async function () {
 		await this._checkQueuedItems;
 	},
 
@@ -260,21 +282,15 @@ Zotero.ExpressionsOfConcern = {
 	 * @param libraryID
 	 * @returns {Promise<boolean>}
 	 */
-	libraryHasExpressionsOfConcern: async function (libraryID) {
-		if (this._expressionsOfConcernByLibrary[libraryID] && this._expressionsOfConcernByLibrary[libraryID].size) {
+	libraryHasExpressionsOfConcern: function (libraryID) {
+		if (
+			this._expressionsOfConcernByLibrary[libraryID]
+			&& this._expressionsOfConcernByLibrary[libraryID].size
+		) {
 			return true;
 		}
 
 		return false;
-	},
-
-	/**
-	 *
-	 * @param itemID
-	 * @returns {Promise<void>}
-	 */
-	hideExpressionOfConcern: async function (itemID) {
-		this._updateEntryFlag(itemID, this.FLAG.HIDDEN);
 	},
 
 	/**
@@ -290,43 +306,21 @@ Zotero.ExpressionsOfConcern = {
 		}
 
 		this._expressionsOfConcernByLibrary[libraryID].add(itemID);
-		this._updateLibraryExpressionsOfConcern(libraryID);
-	},
-
-	/**
-	 *
-	 * @param itemID
-	 * @param libraryID
-	 * @returns {Promise<void>}
-	 * @private
-	 */
-	_removeEntryFromExpressionsOfConcernLibrary: async function (itemID, libraryID) {
-		if (!this._expressionsOfConcernByLibrary[libraryID]) {
-			return;
-		}
-
-		this._expressionsOfConcernByLibrary[libraryID].delete(itemID);
-		this._updateLibraryExpressionsOfConcern(libraryID);
-	},
-
-	/**
-	 *
-	 * @param libraryID
-	 * @private
-	 */
-	_resetExpressionsOfConcernLibrary: function (libraryID) {
-		delete this._librariesWithExpressionsOfConcern[libraryID];
-		this._updateLibraryExpressionsOfConcern(libraryID);
+		await this._updateLibraryExpressionsOfConcern(libraryID);
 	},
 
 	_updateLibraryExpressionsOfConcern: async function (libraryID) {
 		let previous = this._librariesWithExpressionsOfConcern.has(libraryID);
 		let current = this.libraryHasExpressionsOfConcern(libraryID);
 
-		if (Zotero.Libraries.exists(libraryID) && (previous !== current || Zotero.Utilities.Internal.getVirtualCollectionStateForLibrary(libraryID, "expressionsOfConcern"))) {
+		if (
+			Zotero.Libraries.exists(libraryID)
+			&& (previous !== current
+				|| (current && !Zotero.Utilities.Internal.getVirtualCollectionStateForLibrary(libraryID, "expressionOfConcern")))
+		) {
 			let promises = [];
 			for (let zoteroPane of Zotero.getZoteroPanes()) {
-				promises.push(zoteroPane.setVirtual(libraryID, "expressionsOfConcern", current));
+				promises.push(zoteroPane.setVirtual(libraryID, 'expressionOfConcern', current));
 				zoteroPane.hideExpressionsOfConcernBanner();
 			}
 
@@ -343,22 +337,52 @@ Zotero.ExpressionsOfConcern = {
 
 	/**
 	 *
+	 * @param itemID
+	 * @param libraryID
+	 * @returns {Promise<void>}
+	 * @private
+	 */
+	_removeEntryFromExpressionsOfConcernLibrary: async function (itemID, libraryID) {
+		if (!this._expressionsOfConcernByLibrary[libraryID]) {
+			return;
+		}
+
+		this._expressionsOfConcernByLibrary[libraryID].delete(itemID);
+		await this._updateLibraryExpressionsOfConcern(libraryID);
+	},
+
+	/**
+	 *
+	 * @param libraryID
+	 * @private
+	 */
+	_resetExpressionsOfConcernLibrary: function (libraryID) {
+		delete this._librariesWithExpressionsOfConcern[libraryID];
+		this._updateLibraryExpressionsOfConcern(libraryID);
+	},
+
+	/**
+	 *
 	 * @param itemsWithExpressionsOfConcern {[{itemID: string, value: string}]}
 	 * @returns {Promise<void>}
 	 * @private
 	 */
 	_showAlert: async function (itemsWithExpressionsOfConcern) {
 		this._queuedItemIDs.clear();
+
 		// Don't show banner for items in the trash
 		let items = await Zotero.Items.getAsync(itemsWithExpressionsOfConcern);
 		items = items.filter(item => !item.deleted);
+
 		if (!items.length) {
 			return;
 		}
 
 		Zotero.Prefs.set('expressionsOfConcern.recentItems', JSON.stringify(items.map(item => item.id)));
 		this._queuedItemIDs.clear();
+
 		let zoteroPane = Zotero.getActiveZoteroPane();
+
 		if (zoteroPane) {
 			await zoteroPane.showExpressionsOfConcernBanner();
 		}
@@ -384,11 +408,13 @@ Zotero.ExpressionsOfConcern = {
 					this._resetExpressionsOfConcernLibrary(libraryID);
 				}
 			}
+
+			return;
 		}
 
 		if (action === "add") {
 			for (let itemID of ids) {
-				let item = await this.lookupExpressionsOfConcernItem(itemID);
+				let item = await this.lookupUrlForEocs(itemID);
 				if (!item) {
 					return;
 				}
@@ -402,7 +428,7 @@ Zotero.ExpressionsOfConcern = {
 		if (action === "modify") {
 			for (let itemID of ids) {
 				let item = Zotero.Items.get(itemID);
-				let expressionOfConcern = await this.lookupExpressionsOfConcernItem(itemID);
+				let expressionOfConcern = await this.lookupUrlForEocs(itemID);
 				if (!expressionOfConcern) {
 					return;
 				}
@@ -412,18 +438,20 @@ Zotero.ExpressionsOfConcern = {
 				});
 
 				let flag = this._expressionsOfConcern.get(itemID);
-				if (flag !== undefined && (flag === this.FLAG.HIDDEN && item.deleted)) {
-					this._removeEntryFromExpressionsOfConcernLibrary(itemID, item.libraryID);
-				}
-				else {
-					this._addEntryToExpressionsOfConcernLibrary(itemID, item.libraryID);
+				if (flag !== undefined) {
+					if (flag === this.FLAG.HIDDEN || item.deleted) {
+						await this._removeEntryFromExpressionsOfConcernLibrary(item.id, item.libraryID);
+					}
+					else {
+						await this._addEntryToExpressionsOfConcernLibrary(item.id, item.libraryID);
+					}
 				}
 			}
 		}
 
 		if (action === "deleted") {
 			for (let itemID of ids) {
-				this._removeEntry(itemID, extraData[itemID].libraryID);
+				await this._removeEntry(itemID, extraData[itemID].libraryID);
 			}
 		}
 	},
@@ -433,7 +461,7 @@ Zotero.ExpressionsOfConcern = {
 	 * @param item { Zotero.Item } ItItemem which will be checked for expressions of concerns
 	 * @returns {boolean}
 	 */
-	hasExpressionsOfConcern: function (item) {
+	hasExpressionOfConcern: function (item) {
 		let expressionOfConcern = this._expressionsOfConcern.has(item.id);
 
 		if (!expressionOfConcern) {
@@ -446,9 +474,9 @@ Zotero.ExpressionsOfConcern = {
 	/**
 	 * inconsistency in database zotero stores path urls as extensions even though it belongs to the root item
 	 * Query is not working
-	 * @returns {Promise<void>}
+	 * @returns {Promise<{itemID: string, value: string}>}
 	 */
-	lookupExpressionsOfConcernItems: async function () {
+	lookupUrlsForEocs: async function () {
 		const queryString = `SELECT items.itemID, itemDataValues.value
 							FROM items
 							LEFT JOIN itemTypeFields 
@@ -475,7 +503,7 @@ Zotero.ExpressionsOfConcern = {
 	 * Returns the most recent item which has a url from the DB
 	 * @returns {Promise<{itemID: string, value: string}>}
 	 */
-	lookupExpressionsOfConcernItem: async function (itemID) {
+	lookupUrlForEocs: async function (itemID) {
 		const queryString = `SELECT items.itemID, itemDataValues.value
 							FROM items
 							LEFT JOIN itemTypeFields 
@@ -509,6 +537,7 @@ Zotero.ExpressionsOfConcern = {
 			promises.push(Zotero.HTTP.request("GET", item.value, {})
 				.then((response) => {
 					let htmlDoc = response.responseXML;
+					let zoteroItem = Zotero.Items.get(item.itemID);
 
 					if (!htmlDoc) {
 						var parser = Components.classes["@mozilla.org/xmlextras/domparser;1"]
@@ -550,7 +579,7 @@ Zotero.ExpressionsOfConcern = {
 							notices: notices
 						};
 
-						this._addEntry(item.itemID, data);
+						this._addEntry(zoteroItem, data);
 					}
 				})
 				.catch((error) => {
